Make evaluation steps load on scroll; add examples
Some expressions diverge and it's important that the program doesn't crash.
The div in which results are shown now loads more results as the user
scrolls. The other option would have been to implement a timeout, but this
is more educational.

An examples section has been added on the bottom-right. The named
examples are clickable and are entered and evaluated automatically.